LayoutTests/fast/text/international/system-language/composition-expected.txt

PASS target1.offsetWidth is target2.offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
ㄔˉㄔˉㄔˉㄔˉㄔˉㄔˉ
ㄔˉㄔˉㄔˉㄔˉㄔˉㄔˉ
